feat!: Extracted logic from index.js to separate files
