+++ src/utils/registerEventListeners.js
@@ -2,6 +2,8 @@ const logger = require('../services/logger');
 const fs = require('node:fs');
 
 exports.registerEventListeners = (client, player) => {
+    logger.debug('Registering event listeners...');
+
     const eventFolders = fs.readdirSync('./src/events');
     for (const folder of eventFolders) {
         const eventFiles = fs.readdirSync(`./src/events/${folder}`).filter((file) => file.endsWith('.js'));
@@ -10,6 +12,7 @@ exports.registerEventListeners = (client, player) => {
             const event = require(`../events/${folder}/${file}`);
             switch (folder) {
                 case 'client':
+                    logger.debug(`Registering client event listener ${event.name}...`);
                     if (event.once) {
                         client.once(event.name, (...args) => event.execute(...args));
                     } else {
@@ -24,14 +27,17 @@ exports.registerEventListeners = (client, player) => {
                     break;
 
                 case 'interactions':
+                    logger.debug(`Registering interactions event listener ${event.name}...`);
                     client.on(event.name, (...args) => event.execute(...args, { client }));
                     break;
 
                 case 'process':
+                    logger.debug(`Registering process event listener ${event.name}...`);
                     process.on(event.name, (...args) => event.execute(...args));
                     break;
 
                 case 'player':
+                    logger.debug(`Registering player event listener ${event.name}...`);
                     if (
                         !event.isDebug ||
                         process.env.NODE_ENV === 'development' ||
